drf_lms/users/urls.py
from rest_framework.routers import DefaultRouter
from django.urls import path, include
from .views import PaymentViewSet, RegisterView, UserListView, UserDetailView

router = DefaultRouter()
router.register(r'payments', PaymentViewSet, basename='payment')

urlpatterns = [
    path('', include(router.urls)),
    path('register/', RegisterView.as_view(), name='register'),
    path('', UserListView.as_view(), name='user-list'),
    path('<int:pk>/', UserDetailView.as_view(), name='user-detail'),
]